# Conflicts:
#	app/src/main/java/com/makentoshe/androidgithubcitemplate/MainActivity.kt

+++ app/src/main/java/com/makentoshe/androidgithubcitemplate/FieldView.kt
@@ -151,7 +151,7 @@ class FieldView(
     override fun onDraw(canvas: Canvas) {
         fieldSizeY = fieldSizeX * width / height * fieldData.fieldSizeH / fieldData.fieldSizeW
         canvas.apply {
-            val rectWidth: Float = width * fieldSizeX.toFloat() / fieldData.fieldSizeW.toFloat() * zoom
+            val rectWidth: Float = width * fieldSizeX / fieldData.fieldSizeW.toFloat() * zoom
             val matrix = Matrix()
 
             painter.style = Paint.Style.STROKE
